refactor: separate appointment route component
